fix: 默认头像排序等

+++ adminserver/admin-model/src/main/java/com/sbvadmin/model/Permission.java
@@ -87,7 +87,7 @@ public class Permission implements Serializable {
     /**
      * 排序
      */
-    private Byte orderNo;
+    private Integer orderNo;
 
     private LocalDateTime createdAt;
